python-llm-post/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.7.0
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-text-splitters==0.3.3
langsmith==0.2.3
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.4
openai==1.57.4
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
packaging==24.2
pip==24.2
propcache==0.2.1
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.7.0
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
yarl==1.18.3
